refactor(service): 명령어 관련 구조 변경
사용자 입력에 관련된 파일을 패키지로 분리합니다.
구분자를 위한 클래스를 record로 변경합니다.
